Add files via upload
As a Data Analytics Enthusiast, it was so amazing when I got to discover how Orange could be used to do Image Analytics and Predictive Analytics. The secret of using Orange is simply the widgets. I was able to use hierarchical clustering to classify various animals by comparing the cosine distance between the attributes gotten from the image embedding by the DNN. I was also able to predict the political party a president belongs to using both logistic regression and classification tree machine learning algorithm. I got the same prediction but the classification accuracy was higher than the logistic regression method.