Add files via upload
This is a Logistic Regression Model written in Python that predicts the eligibility of customers that requested for loan in real time. The dataset is for Dream Housing Finance company that deals in all home loans. They have presence across all urban semi urban and rural areas. Customer first apply after that company validates the customer eligibility for loan.
The procedure I followed are:
1. I understood the problem statement
2. I generated Hypothesis
3. I got the system ready and loaded the data
4. I understood the data thoroughly
5. I ran Exploratory Data Analysis (EDA) on the data which included Univariate and Bivariate Analysis 
6. I dealt with the Missing values and Outliers properly
7. I built a Logistic Regression model for the problem
8. I used confusion matrix and heat maps to check for classification errors
9. I evaluate the Metrics for the classification problems by scoring the model
10. Feature Engineering