Add files via upload
Gambling is the wagering of money or something of value on an event with an uncertain outcome 
with the primary intent of winning money or material goods. Sports betting is a type of gambling, 
it is the activity of predicting sports results and placing a wager on the outcome. The prevalence 
of sports betting amongst students in the University of Lagos led us to investigate gambling habits 
amongst these students. This investigation involved examining the influence of sports betting on 
the students’ academic performance, behavior and also the dependence of age, gender and religion 
on the students’ sports betting habit. This project involved the use of a stratified sampling method 
where a draft of a questionnaire was developed and a pilot survey was conducted and the 
questionnaire was adjusted based on the responses, the result was used to calculate the sample size 
for the main survey and the main result was inputted into SPSS and analysis was carried out. 
Thereafter, inferences were made based on the result of the analysis gotten. A total of 191 students 
were sampled in the Faculty of Engineering and from the analysis it was seen that a relatively high 
proportion of people in the Faculty of Engineering engaged in sports betting and most of them 
engage in it because of money. It was also discovered that sports betting had a great influence on 
the behavior of the students. And that sports betting does not really affect the academic 
performance of the students.